lib: Extract erofs writing to separate file
This adds a new header lcfs-internal.h which contains things
shared by the basic writer APIs and the erofs specific stuff, and
then moves the erofs code to lcfs-writer-erofs.c.

Some functions in lcfs-writer are now made non-static, but that
is fine as we now hide all non-API symbols.
