+++ libcomposefs/Makefile-lib.am
@@ -4,9 +4,11 @@ libcomposefs_la_SOURCES = \
                         $(COMPOSEFSDIR)/bitrotate.h \
                         $(COMPOSEFSDIR)/hash.c \
                         $(COMPOSEFSDIR)hash.h \
+                        $(COMPOSEFSDIR)/lcfs-internal.h \
                         $(COMPOSEFSDIR)/lcfs-cfs.h \
                         $(COMPOSEFSDIR)/lcfs-fsverity.c \
                         $(COMPOSEFSDIR)/lcfs-fsvertiy.h \
+                        $(COMPOSEFSDIR)/lcfs-writer-erofs.c \
                         $(COMPOSEFSDIR)/lcfs-writer.c \
                         $(COMPOSEFSDIR)/lcfs-writer.h \
                         $(COMPOSEFSDIR)/lcfs-mount.c \
